+++ src/main/java/io/github/yaml/pinterm/PinTermDynamicPluginSupport.java
@@ -1,6 +1,5 @@
 package io.github.yaml.pinterm;
 
-import com.intellij.ide.AppLifecycleListener;
 import com.intellij.ide.plugins.DynamicPluginListener;
 import com.intellij.ide.plugins.IdeaPluginDescriptor;
 import com.intellij.openapi.extensions.PluginId;
@@ -8,22 +7,8 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManager;
 import org.jetbrains.annotations.NotNull;
 
-public final class PinTermDynamicPluginSupport implements DynamicPluginListener, AppLifecycleListener {
+public final class PinTermDynamicPluginSupport implements DynamicPluginListener {
     private static final PluginId PLUGIN_ID = PluginId.getId("io.github.yaml.pinterm");
-
-    @Override
-    public void appStarted() {
-        PinTermIdeDefaultsService.getInstance();
-    }
-
-    @Override
-    public void pluginLoaded(@NotNull IdeaPluginDescriptor pluginDescriptor) {
-        if (!PLUGIN_ID.equals(pluginDescriptor.getPluginId())) {
-            return;
-        }
-
-        PinTermIdeDefaultsService.getInstance();
-    }
 
     @Override
     public void beforePluginUnload(@NotNull IdeaPluginDescriptor pluginDescriptor, boolean isUpdate) {